Separate streaming message functionality into hook
